test: add separate workflows
