+++ tests/missing-tests/test-missing.old.ts
@@ -1,6 +1,6 @@
-import { createMockGitHub, MockGitHub, failureStep } from './utils';
+import { createMockGitHub, MockGitHub, failureStep } from '../utils';
 
-describe('github-action-test-compare', () => {
+describe.skip('github-action-test-compare', () => {
   let mockGitHub: MockGitHub;
 
   beforeEach(async () => {
