+++ tests/test-failure/main/package.json
@@ -1,0 +1,9 @@
+{
+  "scripts": {
+    "test": "vitest --run"
+  },
+  "devDependencies": {
+    "vite-tsconfig-paths": "^4.0.7",
+    "vitest": "^0.29.5"
+  }
+}
