+++ tests/test-failure/main/vitest.config.ts
@@ -1,0 +1,9 @@
+import { defineConfig } from 'vitest/config';
+import viteTsconfigPaths from 'vite-tsconfig-paths';
+
+export default defineConfig({
+  test: {
+    globals: true,
+  },
+  plugins: [viteTsconfigPaths()],
+});
