+++ tests/test-success/main/.github/workflows/test-compare.yml
@@ -1,0 +1,19 @@
+name: pr-non-production
+on:
+  pull_request:
+    types: [opened, reopened, synchronize, labeled]
+
+jobs:
+  run:
+    runs-on: ubuntu-latest
+    permissions:
+      contents: read
+      pull-requests: write
+    steps:
+      - name: Checkout action
+        uses: actions/checkout@v3
+        with:
+          path: testAction
+
+      - name: Test compare
+        uses: ./testAction
